# Don't parallelize coarse motion estimation